websites/requirement.txt
Django==1.11
MySQL-python==1.2.5
Pillow==2.9.0
argparse==1.2.1
django-cors-headers==2.2.0
djangorestframework==3.8.2
pytz==2018.4
uWSGI==2.0.17.1
wsgiref==0.1.2
